Centralize editor snippet response parsing
