+++ packages/tools/nodeParticleEditor/vite.config.ts
@@ -8,6 +8,7 @@ export default defineConfig((_env) => {
         port: parseInt(process.env.NPE_PORT ?? "1345"),
         aliases: {
             "shared-ui-components": path.resolve("../../dev/sharedUiComponents/src"),
+            "@tools/snippet-loader": path.resolve("../../tools/snippetLoader/src"),
         },
         cdnExternals: {
             "@dev/core": "BABYLON",
