+++ packages/tools/nodeRenderGraphEditor/vite.config.ts
@@ -9,6 +9,7 @@ export default defineConfig((_env) => {
         aliases: {
             "shared-ui-components": path.resolve("../../dev/sharedUiComponents/src"),
             gui: path.resolve("../../dev/gui/dist"),
+            "@tools/snippet-loader": path.resolve("../../tools/snippetLoader/src"),
         },
         cdnExternals: {
             "@dev/core": "BABYLON",
